Einige kleinere Korrekturen und Refactoring. Nicht System-Klassen werden jetzt auch geladen. Aber so richtig funktionieren tun sie immer noch nicht.

+++ src/misc/defs.h
@@ -70,8 +70,8 @@ typedef unsigned long long uint64_t;
 typedef long long int64_t;
 typedef unsigned long uint32_t;
 typedef long int32_t;
-typedef unsigned int uint16_t;
-typedef int int16_t;
+typedef unsigned short uint16_t;
+typedef short int16_t;
 typedef unsigned char uint8_t;
 typedef char int8_t;
 #endif
